Code/requirements.txt
caldav==1.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
icalendar==5.0.4
idna==3.4
lxml==4.9.2
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
recurring-ical-events==2.0.2
requests==2.28.2
six==1.16.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
vobject==0.9.6.1
x-wr-timezone==0.0.5
cryptocode==0.1
pydantic==1.10.7
email_validator==2.0.0.post2
sqlalchemy==1.4.39
python-dotenv=1.0.0
bcrypt==4.0.1